soclib/requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.1
cryptography==42.0.5
Django==5.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
